REmoved transfer controller as it is easier to use the employercommitmentsController. And added the razer view
