+++ src/SFA.DAS.EAS.Web/SFA.DAS.EmployerCommitments.Web.csproj
@@ -441,7 +441,6 @@
     <Compile Include="Authentication\OwinWrapper.cs" />
     <Compile Include="Controllers\AccessDeniedController.cs" />
     <Compile Include="Controllers\DataLockController.cs" />
-    <Compile Include="Controllers\TransferController.cs" />
     <Compile Include="Controllers\ServiceController.cs" />
     <Compile Include="Decorators\AccountApiClientDecorator.cs" />
     <Compile Include="DependencyResolution\StructureMapValidatorFactory.cs" />
@@ -660,7 +659,6 @@
       <Generator>RazorGenerator</Generator>
       <LastGenOutput>_CoreDetail.generated.cs</LastGenOutput>
     </Content>
-    <Content Include="Views\Transfer\TransferDetails.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="comt-assets\css\external-links\external-link-24x24.png" />
@@ -1135,6 +1133,7 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="Views\EmployerCommitments\SelectTransferConnection.cshtml" />
+    <Content Include="Views\EmployerCommitments\TransferDetails.cshtml" />
     <None Include="Views\razorgenerator.directives" />
   </ItemGroup>
   <ItemGroup>
